# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Zimbabwe Skill Verification Platform API
  version: 1.0.0
  description: API for institution-managed credentials, public verification, blockchain evidence, auditability, and analytics. Examples contain placeholders only. The uploaded PDF is authoritative; a generated PDF is a presentation copy. Credentials become active only after blockchain confirmation, and revoked credentials remain publicly verifiable.
  contact:
    name: API Support
    email: api-support@example.org
  license:
    name: ISC
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://api.example.org
    description: Production placeholder
...
paths:
...
  /api/auth/reset-password:
    post: { tags: [Authentication], summary: Consume a single-use password reset token, description: Does not establish a session. Successful use revokes existing sessions and clears account lock state., operationId: resetPassword, requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ResetPasswordInput' } } } }, responses: { '200': { $ref: '#/components/responses/Success' }, '400': { $ref: '#/components/responses/ValidationError' }, '409': { $ref: '#/components/responses/Conflict' }, '429': { $ref: '#/components/responses/RateLimited' } } }
  /api/auth/profile:
    get: { tags: [Authentication], summary: Current safe user profile, operationId: profile, security: [ { BearerAuth: [] } ], responses: { '200': { $ref: '#/components/responses/Success' }, '401': { $ref: '#/components/responses/AuthRequired' } } }
    patch: { tags: [Authentication], summary: Update current name or email, operationId: updateProfile, security: [ { BearerAuth: [] } ], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ProfileInput' } } } }, responses: { '200': { $ref: '#/components/responses/Success' }, '400': { $ref: '#/components/responses/ValidationError' }, '401': { $ref: '#/components/responses/AuthRequired' }, '409': { $ref: '#/components/responses/Conflict' } } }
  /api/auth/change-password:
    post: { tags: [Authentication], summary: Change current password and invalidate existing tokens, operationId: changePassword, security: [ { BearerAuth: [] } ], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ChangePasswordInput' } } } }, responses: { '200': { $ref: '#/components/responses/Success' }, '400': { $ref: '#/components/responses/ValidationError' }, '401': { $ref: '#/components/responses/AuthRequired' }, '409': { $ref: '#/components/responses/Conflict' } } }
...
  /api/users:
    get: { tags: [Users], summary: List users with server pagination, operationId: listUsers, security: [ { BearerAuth: [] } ], x-roles: [super_admin, institution_admin], parameters: [ { $ref: '#/components/parameters/Page' }, { $ref: '#/components/parameters/Limit' }, { $ref: '#/components/parameters/Search' }, { $ref: '#/components/parameters/SortOrder' }, { name: sortBy, in: query, schema: { type: string, enum: [created_at, updated_at, full_name, email, role, is_active] } }, { name: role, in: query, schema: { $ref: '#/components/schemas/Role' } }, { name: status, in: query, schema: { type: string, enum: [active, inactive, 'true', 'false'] } }, { name: institutionId, in: query, schema: { type: string, format: uuid } } ], responses: { '200': { $ref: '#/components/responses/Paginated' }, '400': { $ref: '#/components/responses/ValidationError' }, '401': { $ref: '#/components/responses/AuthRequired' }, '403': { $ref: '#/components/responses/AccessDenied' } } }
    post: { tags: [Users], summary: Create a managed user, operationId: createUser, security: [ { BearerAuth: [] } ], x-roles: [super_admin, institution_admin], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/UserInput' } } } }, responses: { '201': { $ref: '#/components/responses/Created' }, '400': { $ref: '#/components/responses/ValidationError' }, '403': { $ref: '#/components/responses/AccessDenied' }, '409': { $ref: '#/components/responses/Conflict' }, '422': { $ref: '#/components/responses/ValidationError' } } }
...
  /api/users/{id}/unlock:
    post: { tags: [Users], summary: Clear an authorized user's account lock, description: Clears failed-login and lock state without changing the password or creating a session., operationId: unlockUser, security: [ { BearerAuth: [] } ], x-roles: [super_admin, institution_admin], parameters: [ { $ref: '#/components/parameters/Id' } ], requestBody: { content: { application/json: { schema: { type: object, additionalProperties: false } } } }, responses: { '200': { $ref: '#/components/responses/Success' }, '403': { $ref: '#/components/responses/AccessDenied' }, '404': { $ref: '#/components/responses/NotFound' } } }
  /api/users/{id}/require-password-change:
    post: { tags: [Users], summary: Require a password change and invalidate sessions, operationId: requireUserPasswordChange, security: [ { BearerAuth: [] } ], x-roles: [super_admin, institution_admin], parameters: [ { $ref: '#/components/parameters/Id' } ], requestBody: { content: { application/json: { schema: { type: object, additionalProperties: false } } } }, responses: { '200': { $ref: '#/components/responses/Success' }, '403': { $ref: '#/components/responses/AccessDenied' }, '404': { $ref: '#/components/responses/NotFound' } } }
...
components:
  securitySchemes:
    BearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
    CookieAuth: { type: apiKey, in: cookie, name: verifyzw_session, description: HttpOnly browser session cookie; cookie-authenticated unsafe requests also require X-CSRF-Token. }
  parameters:
    Id: { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    Page: { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
    Limit: { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
    Search: { name: search, in: query, schema: { type: string, maxLength: 200 } }
    SortOrder: { name: sortOrder, in: query, schema: { type: string, enum: [asc, desc], default: desc } }
...
  schemas:
    Role: { type: string, enum: [super_admin, institution_admin, issuer, verifier] }
...
    StandardError: { type: object, required: [success, message], properties: { success: { type: boolean, enum: [false] }, code: { type: string, example: ACCESS_DENIED }, message: { type: string, example: You do not have permission to perform this action. }, requestId: { type: string, format: uuid, nullable: true } } }
    ValidationError: { allOf: [ { $ref: '#/components/schemas/StandardError' }, { type: object, properties: { code: { type: string, enum: [VALIDATION_ERROR] }, errors: { type: array, items: { type: object, required: [field, message], properties: { field: { type: string, example: email }, message: { type: string, example: A valid email address is required. } } } } } } ] }
    Pagination: { type: object, required: [page, limit, total, totalPages, hasNextPage, hasPreviousPage], properties: { page: { type: integer }, limit: { type: integer }, total: { type: integer }, totalPages: { type: integer }, hasNextPage: { type: boolean }, hasPreviousPage: { type: boolean } } }
...
    ResetPasswordInput: { type: object, required: [token, password, confirmPassword], properties: { token: { type: string, minLength: 43, example: '<PASSWORD_RESET_TOKEN>' }, password: { type: string, format: password, minLength: 12, example: '<NEW_STRONG_PASSWORD>' }, confirmPassword: { type: string, format: password, minLength: 12, example: '<NEW_STRONG_PASSWORD>' } }, additionalProperties: false }
...
    ProfileInput: { type: object, minProperties: 1, properties: { fullName: { type: string, minLength: 2, maxLength: 150 }, email: { type: string, format: email } }, additionalProperties: false }
    ChangePasswordInput: { type: object, required: [currentPassword, newPassword, confirmPassword], properties: { currentPassword: { type: string, format: password, example: '<CURRENT_PASSWORD>' }, newPassword: { type: string, format: password, example: '<NEW_STRONG_PASSWORD>' }, confirmPassword: { type: string, format: password, example: '<NEW_STRONG_PASSWORD>' } }, additionalProperties: false }
...
    UserInput: { type: object, required: [fullName, email, role], properties: { fullName: { type: string, minLength: 2, maxLength: 150 }, email: { type: string, format: email }, role: { $ref: '#/components/schemas/Role' }, institutionId: { type: string, format: uuid, nullable: true }, isActive: { type: boolean } }, additionalProperties: false }
...
  responses:
    Success: { description: Successful operation, content: { application/json: { schema: { type: object, required: [success], properties: { success: { type: boolean, enum: [true] } }, additionalProperties: true } } } }
    Created: { description: Resource created, content: { application/json: { schema: { type: object, required: [success], properties: { success: { type: boolean, enum: [true] } }, additionalProperties: true } } } }
    Paginated: { description: Paginated result; non-super-admin results are institution scoped where applicable, content: { application/json: { schema: { type: object, required: [success, pagination], properties: { success: { type: boolean, enum: [true] }, pagination: { $ref: '#/components/schemas/Pagination' } }, additionalProperties: true } } } }
...
    ValidationError: { description: Request validation failed, content: { application/json: { schema: { $ref: '#/components/schemas/ValidationError' } } } }
    AuthRequired: { description: Missing, expired, stale, or invalid bearer token, content: { application/json: { schema: { $ref: '#/components/schemas/StandardError' } } } }
...
    AccessDenied: { description: Authenticated actor lacks permission or institution scope, content: { application/json: { schema: { $ref: '#/components/schemas/StandardError' } } } }
...
    NotFound: { description: Resource not found or feature disabled, content: { application/json: { schema: { $ref: '#/components/schemas/StandardError' } } } }
    Conflict: { description: Duplicate or conflicting resource state, content: { application/json: { schema: { $ref: '#/components/schemas/StandardError' } } } }
...
    RateLimited: { description: Request rate limit exceeded, headers: { Retry-After: { schema: { type: integer } } }, content: { application/json: { schema: { $ref: '#/components/schemas/StandardError' } } } }
...